+++ app/src/main/java/com/example/animemoi_app/common/history/HistoryCard.kt
@@ -1,16 +1,7 @@
 package com.example.animemoi_app.common.history
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -25,10 +16,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.animemoi_app.data.ComicData
 import com.example.animemoi_app.model.Comic
 
 @Composable
@@ -38,7 +27,7 @@ fun HistoryCard(
     Card(
         modifier = Modifier
             .fillMaxWidth()
-            .height(100.dp)
+            .height(90.dp)
             .padding(horizontal = 15.dp),
         elevation = CardDefaults.cardElevation(10.dp),
         colors = CardDefaults.cardColors(containerColor = Color.Transparent),
@@ -57,7 +46,7 @@ fun HistoryCard(
                 modifier = Modifier
                     .padding(start = 5.dp)
                     .fillMaxHeight(),
-                verticalArrangement = Arrangement.spacedBy(20.dp)
+                verticalArrangement = Arrangement.SpaceBetween
             ) {
                 Text(
                     text = stringResource(id = comic.stringResourceId),
@@ -69,14 +58,14 @@ fun HistoryCard(
                     overflow = TextOverflow.Ellipsis
                 )
                 Text(
-                    text = stringResource(id = comic.statusChapter),
+                    text = stringResource(id = comic.status),
                     color = Color.White,
                     fontWeight = FontWeight.Light,
                     fontFamily = FontFamily.Monospace,
                     fontSize = 13.sp
                 )
                 Text(
-                    text = "Tình trạng: " + stringResource(id = comic.statusComic),
+                    text = "Tình trạng: " + stringResource(id = comic.status),
                     color = Color.White,
                     fontWeight = FontWeight.Light,
                     fontFamily = FontFamily.Monospace,
@@ -85,25 +74,4 @@ fun HistoryCard(
             }
         }
     }
-}
-
-@Composable
-fun GridHistoryCard() {
-    LazyVerticalGrid(
-        columns = GridCells.Fixed(1),
-        verticalArrangement = Arrangement.spacedBy(20.dp),
-        //horizontalArrangement = Arrangement.spacedBy(8.dp),
-        modifier = Modifier.padding(horizontal = 20.dp)
-    ) {
-        val comics = ComicData().loadComicCard()
-        items(comics) { comic ->
-            HistoryCard(comic)
-        }
-    }
-}
-
-@Preview
-@Composable
-fun HistoryCardPreview() {
-    GridHistoryCard()
 }
