+++ app/src/main/res/values/chapter.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="newest">Chương mới nhất</string>
+    <string name="newest">Chương %d</string>
 </resources>
